contact-keeper/backend/requirements.txt
Flask==2.3.3
phonenumbers==8.13.55
psycopg2-binary==2.9.9
python-dotenv==1.0.0

